Change energy to int instead of double
- Save beam energy as int and not double (framework is not made for 
precise beam energy studies).

+++ source/prew/include/Data/DistrInfo.h
@@ -12,7 +12,7 @@ namespace Data {
     std::string m_process_name {};  // Name of process
     std::string m_distr_name {};    // Name of distribution
     std::string m_pol_config {};    // Name of polarisation setting
-    double      m_energy {};        // Center of mass energy
+    int         m_energy {};        // Center of mass energy
   };
   
 }
